Small refactoring of tests
