+++ .travis.yml
@@ -19,6 +19,7 @@ matrix:
   fast_finish: true
   include:
     - env: OGGM_TEST_ENV=prepro MPL=
+    - env: OGGM_TEST_ENV=numerics MPL=
     - env: OGGM_TEST_ENV=models MPL=
     - env: OGGM_TEST_ENV=workflow MPL=--mpl
     - env: OGGM_TEST_ENV=graphics MPL=--mpl
